fix(i18n): 拆 language_pref.rs / graph.rs / routes.rs 三个 god file (#155)
sentrux 的 god_file 判据是 loc>800 || (functions>25 && loc>400)
（sentrux_gate.rs:794）。#155 這一支把 god_file_count 从 main@9d00329 的 33
顶到了 34（files 285→288），三个触发点：

- language_pref.rs（新文件，423 loc / 27 fn，命中第二条件）
- graph.rs（564→596 loc，因为加了 issue #155 验收要求的 zh/en 产物逐字节
  相等测试，fn 25→26，刚好从边界之下越过第二条件）
- cli/command_catalog/routes.rs（新增 language 的 legacy_route! 条目后
  loc 800→801，单独命中 loc>800）——这一个之前没测出来，是本轮复核时
  用 sentrux_gate.rs 里 collect_files/measure_file 的原样逻辑手写 Python
  复现脚本扫出来的。

统一按这仓库既有的目录模块惯例拆（change_risk/、change_agenda/ 的先例）：
mod.rs 留主体，tests.rs 单独放测试。language_pref.rs/graph.rs 两个是
「测试挪出去」，routes.rs 没有内嵌测试模块，改成挪类型定义
（RawRoute/LegacyRoute/VersionRoute/CommandRoute 连同 CommandRoute 枚举）
到新的 routes/types.rs——可见性从 pub(super) 改成显式的
pub(in crate::cli::command_catalog)，因为 super 在新位置只到 routes 这
一级，显式路径保证和原来（可见到 cli::command_catalog 及其 tests 子模块）
完全等价，不多不少。

graph.rs 拆分牵出两处连带修复：
- doctor_bootstrap/mod.rs 的 graphProvider.sourceFound 硬编码检查
  src/graph.rs，现在文件搬到 src/graph/mod.rs，不修的话这项检查会从
  true 变 false（doctor 报告「内部 graph provider 缺失」这种假象）。
- builtin_provider_evidence.rs 里 include_bytes!("graph.rs") 同理改成
  include_bytes!("graph/mod.rs")，这是它给 graph-builtin provider 生成
  运行时 digest 用的，路径不对会直接编译失败（找不到文件）。
- 两处 Rust 测试（doctor_bootstrap_cli.rs、doctor_envelope.rs）和两处
  PowerShell 测试（test-regression-fixes.ps1 的 New-DoctorScratchRoot）
  在 scratch checkout 里手搭一份假 graph.rs 来喂 sourceFound 检查，同步
  改到 src/graph/mod.rs，否则这几个测试会因为检查的是真实行为而报错，
  不是因为它们错，是因为它们喂错了路径。
- orchestration/integrations.json、orchestration/internalization/
  graph.json 两处仍写着 crates/code-intel-cli/src/graph.rs 的 sha256
  digest pin：repin --write 只认领「同一路径、内容变了」的场景，认不出
  「路径本身变了」（deleted/renamed source path），会诚实地把旧 hash
  标成 orphanedPins 而不是悄悄当作 clean——这两处手动改成新路径 +
  graph/mod.rs 的真实 sha256。

验证：
- cargo fmt --check：干净
- cargo test -p code-intel：3017/3017（和拆分前一致，移动测试模块不改
  变用例数）
- ./target/debug/code-intel.exe sentrux scan .：god_file_count 回到 33，
  files=289；另外拿 9d00329 在 C:\ci-baseline-9d00329（短路径，避免
  Windows MAX_PATH 链接失败）干净编译复现同一份 scan，god 文件集合逐个
  比对和 baseline 完全一致（33/33 同名同批，非仅数字巧合）
- repin --write --json 后再跑一次不带 --write 确认 stalePins:[]，然后
  重新跑一遍 cargo test 才当数——上一轮 cherry-pick 冲突就在这一步被
  repin 的假 clean 骗过一次（见 #133 证据，另行汇报）

Refs #155 #133
