+++ crates/code-intel-cli/src/cli/command_catalog/routes/mod.rs
@@ -7,6 +7,9 @@ use crate::cli::help_contract::{HELP_ALIASES, HELP_COMMAND};
 
 mod edit_routes;
 mod run_routes;
+mod types;
+
+pub(super) use types::{CommandRoute, LegacyRoute, RawRoute, VersionRoute};
 
 macro_rules! command_contract {
     ($stability:ident, $controller:ident, Conditional($condition:ident), $effects:expr, $output:expr, $exit:expr, $retirement:literal) => {
@@ -67,36 +70,6 @@ macro_rules! exits {
     ($($code:literal),+ $(,)?) => {
         ExitContract::Exact(&[$($code),+])
     };
-}
-
-pub(super) struct RawRoute {
-    pub(super) command: &'static str,
-    pub(super) subcommand: Option<&'static str>,
-    pub(super) argument_offset: usize,
-    pub(super) id: CompatibilityRoute,
-    #[allow(dead_code)]
-    pub(super) contract: CommandContract,
-}
-
-pub(super) struct LegacyRoute {
-    pub(super) command: &'static str,
-    pub(super) aliases: &'static [&'static str],
-    pub(super) id: LegacyRouteId,
-    #[allow(dead_code)]
-    pub(super) contract: CommandContract,
-}
-
-pub(super) struct VersionRoute {
-    pub(super) command: &'static str,
-    pub(super) aliases: &'static [&'static str],
-    pub(super) contract: CommandContract,
-}
-
-pub(super) enum CommandRoute {
-    Version(VersionRoute),
-    Primary(CommandContract),
-    Raw(RawRoute),
-    Legacy(LegacyRoute),
 }
 
 macro_rules! raw_route {
