+++ docs/architecture/reference-capability-map.md
@@ -20,7 +20,7 @@ This map classifies the active integrations and design references named by the r
 | inventory.rg | Exact file inventory and text-search surface | Observed repository evidence | Normal runner invokes `rg` and writes `files.txt` plus the `rg file inventory` step | Pipeline owns invocation/artifact contract; upstream owns `rg` | `active-required` | Low CLI coupling, currently embedded in the main runner | Inventory is bounded by scope/exclusions and snapshot; absence from a search is not proof of semantic absence outside that scope | Replace the executable behind an inventory adapter while preserving exact inventory artifacts and failure semantics | `orchestration/integrations.json` integration id `inventory.rg`; `legacy/run-code-intel.ps1` step name `rg file inventory` |
 | Git | Working-tree state, repository identity, HEAD/history and recent-commit context | Observed version-control evidence | `git status`, `git rev-parse`, `git log` from runner/adapters | Pipeline owns interpretation; Git owns repository protocol/CLI | `drift-unregistered` | Medium: multiple scripts call Git directly and CodeNexus-lite uses history context | Dirty state, HEAD and worktree policy must be explicit in Snapshot Identity; Git commands are read-only unless a separately authorized effect declares mutation | Register a version-control evidence adapter and centralize snapshot semantics; alternate VCS providers may conform to the same port | `legacy/run-code-intel.ps1` step name `git status`; `legacy/Invoke-CodeNexusLite.ps1`; absent from `orchestration/integrations.json` |
 | Repowise | Semantic index, state, scoped shadow workspace, optional generated docs | Observed evidence and semantic-memory artifacts; provider-generated prose remains non-authoritative | `legacy/Invoke-ScopedRepowise.ps1`; normal runner invokes it; Rust provider routes exist for status/index/docs | Pipeline owns adapter and egress contract; upstream owns CLI/provider internals | `active-required` | Medium/high: CLI plus Python/package behavior is still wrapped and patched locally | Scope inventory and frozen provider payload must be validated before network/provider execution; secrets remain outside artifacts | Preserve `.repowise/state.json`/`wiki.db` and provider-port semantics; replace CLI behind adapter or complete selective internalization | `orchestration/integrations.json` integration id `memory.repowise`; `docs/artifact-data-contract.md` section “Scoped Repowise Egress Contract”; `crates/code-intel-cli/src/providers.rs`; `overlays/repowise/README.md` |
-| Understand-compatible internal graph | Architecture graph state, freshness, compatible graph artifact | Observed structural evidence promoted only after snapshot/freshness validation | `code-intel graph --repo ... --write --json`; normal run consumes `.understand-anything/knowledge-graph.json` | Pipeline | `active-required` | Low/medium: internal Rust producer, but artifact name/schema retains compatibility | Graph freshness and snapshot identity must match the examined repository; stale presence is not success | Any graph provider may replace implementation if it emits the Pipeline-owned compatible artifact through the provider port | `orchestration/integrations.json` integration id `graph.code-intel-understand`; `docs/integration-orchestration.md` section “Understand Direction”; `crates/code-intel-cli/src/graph.rs` |
+| Understand-compatible internal graph | Architecture graph state, freshness, compatible graph artifact | Observed structural evidence promoted only after snapshot/freshness validation | `code-intel graph --repo ... --write --json`; normal run consumes `.understand-anything/knowledge-graph.json` | Pipeline | `active-required` | Low/medium: internal Rust producer, but artifact name/schema retains compatibility | Graph freshness and snapshot identity must match the examined repository; stale presence is not success | Any graph provider may replace implementation if it emits the Pipeline-owned compatible artifact through the provider port | `orchestration/integrations.json` integration id `graph.code-intel-understand`; `docs/integration-orchestration.md` section “Understand Direction”; `crates/code-intel-cli/src/graph/mod.rs` |
 | Understand Anything external | Rich Agent-side architecture graph refresh | Observed evidence from an external/interactive provider | `/understand <repo> --language zh [--full]` only as fallback or explicit richer pass | Upstream provider; Pipeline owns admission of its artifact | `compatibility-fallback` | Low at runtime, medium artifact compatibility | External command is non-deterministic/host-dependent; output must pass the same snapshot/freshness contract | Remove fallback when internal or another non-interactive provider reaches required conformance | `orchestration/integrations.json` integration id `graph.understand-external`; `docs/integration-orchestration.md` section “Understand Direction”; `skills/code-intel-pipeline/SKILL.md` graph-missing workflow |
 | Sentrux | Structural scan, rules, regression gate, DSM, test gaps, evolution, what-if | Observed structural evidence plus derived governance models/gate results | `legacy/Invoke-SentruxAgentTool.ps1`; Rust `code-intel sentrux ...`; external binary through repo shim with lite-core fallback | Pipeline owns adapter, normalization, rules and lite fallback; external binary is accelerator | `active-required` | High today: large PowerShell wrapper; external CLI format is normalized locally | Raw CLI status/exit code is not authoritative until normalized; governed scope, exclusions, baseline and rules are explicit | Migrate operations atom-by-atom behind Rust/provider contracts; retain lite survival capability; external accelerator remains replaceable | `orchestration/integrations.json` integration id `structure.sentrux`; `docs/integration-orchestration.md` section “Sentrux Direction”; `crates/code-intel-cli/src/sentrux.rs`; `legacy/tools/sentrux-shim/sentrux-lite-core.ps1` |
 | tree-sitter-v | V-language parsing grammar for Sentrux | Third-party parser artifact used to create structural evidence | `legacy/Install-SentruxVlangOverlay.ps1`; installed as Sentrux `vlang` plugin | Pipeline owns overlay/ABI alias and installer; upstream owns grammar | `active-optional` | Medium: compiled grammar ABI and Sentrux plugin contract | Compiled artifact provenance/ABI must be verified; current repository records source URL and MIT license but no pinned source revision | Drop overlay when upstream Sentrux ships a conforming V plugin, or rebuild from a pinned reviewed revision | `overlays/sentrux/vlang/plugin.toml:11`; `overlays/sentrux/vlang/plugin.toml:14`; `overlays/sentrux/vlang/README.md:7`; `legacy/Install-SentruxVlangOverlay.ps1` |
