+++ orchestration/integrations.json
@@ -286,7 +286,7 @@
           "version": "1.0.0",
           "toolchainDigests": [
             "c4379b9fa2d7cbc6b39bd08e200cc2ed685da0eb30e99f86facc49e8040fd4d9",
-            "02c5ca1015180a02fb16d0d35c6ba62ceee8516124db03eb49458ee200a01ca6",
+            "b72c4d641441f709f12a7775af848610f4e77eac0c75e18b4b46d75424497e42",
             "7db9e80857f3c65aada25041986cb5cb3b50b95fa69de794d55ad38fd4a8d8d4",
             "1ea59e6fc535572f1bf6ec93fd3ed0857c6b1dcc03db8370c060eec4741da415",
             "7c5cabf7a84fdc2fad9778d45217a66dd3e7f4730ff87012bebc85545b93bf9e",
@@ -533,8 +533,8 @@
         "id": "provider.graph-builtin.compat",
         "version": "1.0.0",
         "toolchainDigests": [
-              "4dbd7269d126da4428d6de7c017b1e202c329a7f21c87d2ef1d58a678653bd33",
-              "d8949280d01ec0af8526f94b77ae9c896958f0013d8dbbe56ecf07b3f0247d0f",
+              "300d911f99d9b026606a0d3118f24ed33dad025b0e0212e3ed84705e80760a90",
+              "59ef3070547d4af845629a160fdf37e8e1b3edbf9883f5ff882b99839e62c82d",
               "68a4f55b76080575721cde3ee62fd746999c6df8e79303b809baecf5d2514338",
               "52644a812174988ede91d98ddfec63c6a91f8478277d7bf74c73f106dd0f776b",
               "98ccc64478b2c61bfd7af741ea1f8ee01a88094065c0f025700e8110b525ef26"
@@ -549,7 +549,7 @@
       "algorithm": "sha256",
       "inputs": [
         "crates/code-intel-cli/src/builtin_provider_evidence.rs",
-        "crates/code-intel-cli/src/graph.rs",
+        "crates/code-intel-cli/src/graph/mod.rs",
         "crates/code-intel-cli/src/graph_adapter.rs",
         "crates/code-intel-cli/src/admissibility.rs",
         "crates/code-intel-cli/src/adapter_contract.rs"
@@ -585,7 +585,7 @@
         "id": "provider.sentrux-builtin.compat",
         "version": "1.0.0",
         "toolchainDigests": [
-              "4dbd7269d126da4428d6de7c017b1e202c329a7f21c87d2ef1d58a678653bd33",
+              "300d911f99d9b026606a0d3118f24ed33dad025b0e0212e3ed84705e80760a90",
               "29ad124d984f6ab0756bdc7c0b523a84dfada2a655fe0d282ce681165f6bae4f",
               "e40f970f48d3ddbb5ff096068988d8c45daacdf4a1a6797fa54e15e40357516c",
               "52644a812174988ede91d98ddfec63c6a91f8478277d7bf74c73f106dd0f776b",
